# Cleanup flush logic